Reworked around console producer limitations.
Since the Avro console producer won't produce primitive key types, we're faced with a tradeoff of the complexity of a second Avro type for they key, or with having to repartition the Movie and Rating topics to make them joinable. I don't like either solution, but I dislike this one (rekeying) the least.