Tumbling Windows in Kafka Streams (#55)
* Exclude Eclipse project files.

Yes, I actually found some lying around in a KStreams code sample. Weird, amirite?

* recipe WIP

* Incremental progress against window recipe

* De-Drogalisified the KSQL images

* Code that compiles, at least! Which is saying something

* Imagine code that almost works. That is this commit.

The current stream processor ends in a call to `foreach(printf(...))`, which is silly. When I attempt to `to()` the results to a topic and then inspect that topic in any reasonable way, I see blank lines. So, still digging.

* I think this is working code

* Getting tests cleaned up

* Sort of a test. Doesn't pass yet.

* Hacking on unit test

* Completed the narrative for tumbling windows

* Experimenting with more recent time stamps

* Made topic retention much longer

* Converted input timestamps to ISO-8601

This involved changing the TimestampExtractor to convert from that format, and changing the input data as well. Now it's obvious what the event time of the input data is just by looking at it. Epoch timestamps were too hard to think about.

* Working recipe, passing unit test

* Fuller  narrative

* Tightened up CI harnesses

* Progress on harnesses

* Dev harness runs to completion

* What's a broken Jekyll build between friends?

* Fixed Jekyll build

* dev harness works

* This commit features working Jekyll as well

* A few fixes.

* Tried the harness on my own machine, and couldn't run without a directly gradlew invocation.

* Fixed timestamp processing and unit test for same

* Also fixed formatting of input data to match the fixed date parser

* Fixed a mistake in the narrative that referred to the wrong config parameter

* Removed some unneeded imports