update rekey join tutorial inline with changes to ksqlDB master
This is a larger change than the other updates as the recent changes to key handling made the existing tutorial not really work. It was relying on the fact that a stream without a key column has `ROWKEY` set to `null` if you issue a push query.

Recent changes mean that a stream without an explicit key column have not key column, i.e. no `ROWKEY`, so I've updated the tutorial to use `DESCRIBE` to see the schema of the data and ``PRINT` to view the raw data in Kafka.