update tumbling windows tutorial inline with changes to ksqlDB master
This includes adding a new SMT to remove the `city_id` from the value. This is needed if we want to name the primary key in ksqlDB `city_id`. Without it, there would be a name clash with the PK and the `city_id` column in the value schema.

Also added some details of _what_ the SMTs are doing, as its not necessarily clear to users.