MINOR: Replace deprectated 'TIMESTAMPTOSTRING' with 'FORMAT_TIMESTAMP' (#1341)
Updated insert statements to use 'UNIX_TIMESTAMP()' to generate dates

This PR proposes to solve the "too old" date issue by using UNIX_TIMESTAMP() to generate dates for inserting the initial data for the tutorial. Additionally, the tutorial uses a deprecated method TIMESTAMPTOSTRING, so it's been replaced with the recommendedFORMAT_TIMESTAMP function.